dataset/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
appnope=0.1.2=py39h2804cbe_2
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39h5161555_1
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=22.1.0=pyhd8ed1ab_0
bleach=4.1.0=pyhd8ed1ab_0
brotlipy=0.7.0=py39h5161555_1003
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2021.10.8=h4653dfc_0
cairo=1.16.0=hb99c34e_1009
certifi=2021.10.8=py39h2804cbe_1
cffi=1.15.0=py39h52b1de0_0
charset-normalizer=2.0.11=pyhd8ed1ab_0
click=8.0.3=py39h2804cbe_1
cryptography=36.0.1=py39hfb8cd70_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.5.1=py39hfb83b0d_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.2=pyhd8ed1ab_0
ffmpeg=4.3.2=ha8417bf_1
flit-core=3.6.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.94=heb65262_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=h17b34a0_1
fribidi=1.0.10=h27ca646_0
future=0.18.2=py39h2804cbe_4
gettext=0.19.8.1=h049c9fb_1008
giflib=5.2.1=h27ca646_2
gmp=6.2.1=h9f76cd9_0
gnutls=3.6.13=h706517b_1
graphite2=1.3.13=h9f76cd9_1001
harfbuzz=3.3.1=h6efa254_0
hdf5=1.12.1=nompi_had0e5e0_103
icu=69.1=hbdafb3b_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.10.1=py39h2804cbe_0
importlib_resources=5.4.0=pyhd8ed1ab_0
ipykernel=6.8.0=py39h32adebf_0
ipython=8.0.1=py39h2804cbe_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.5=pyhd8ed1ab_0
jasper=2.0.33=hba35424_0
jbig=2.1=h3422bc3_2003
jedi=0.18.1=py39h2804cbe_0
jinja2=3.0.3=pyhd8ed1ab_0
jpeg=9e=h3422bc3_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter=1.0.0=py39h2804cbe_7
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_console=6.4.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py39h2804cbe_1
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
krb5=1.19.2=hd92b7a7_3
lame=3.100=h27ca646_1001
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=13_osxarm64_openblas
libcblas=3.9.0=13_osxarm64_openblas
libcurl=7.81.0=h8fe1914_0
libcxx=12.0.1=h168391b_1
libdeflate=1.8=h3422bc3_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libglib=2.70.2=h67e64d8_2
libiconv=1.16=h642e427_0
liblapack=3.9.0=13_osxarm64_openblas
liblapacke=3.9.0=13_osxarm64_openblas
libnghttp2=1.46.0=he4cd7f6_0
libopenblas=0.3.18=openmp_h5dd58f0_0
libopencv=4.5.5=py39hb8da993_0
libpng=1.6.37=hf7e6567_2
libprotobuf=3.19.4=hccf11d3_0
libsodium=1.0.18=h27ca646_1
libssh2=1.10.0=hb80f160_2
libtiff=4.3.0=h74060c4_2
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.12=hedbfbf4_1
libzlib=1.2.11=hee7b306_1013
llvm-openmp=12.0.1=hf3c4609_1
lz4-c=1.9.3=hbdafb3b_1
markupsafe=2.0.1=py39h5161555_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39h5161555_1005
mypy_extensions=0.4.3=py39h2804cbe_4
nbclient=0.5.10=pyhd8ed1ab_1
nbconvert=6.4.1=py39h2804cbe_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.3=hc470f4d_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.6=hc6a1b29_0
ninja=1.10.2=hc021e02_1
notebook=6.4.8=pyha770c72_0
numpy=1.22.2=py39h61a45d2_0
opencv=4.5.5=py39hdf13c20_0
openh264=2.1.1=habe5f53_0
openjpeg=2.4.0=h062765e_1
openssl=1.1.1l=h3422bc3_0
packaging=21.3=pyhd8ed1ab_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.9.0=pyhd8ed1ab_0
pcre=8.45=hbdafb3b_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=py39hcb29f89_0
pip=22.0.3=pyhd8ed1ab_0
pixman=0.40.0=h27ca646_0
platformdirs=2.4.1=pyhd8ed1ab_1
prometheus_client=0.13.1=pyhd8ed1ab_0
prompt-toolkit=3.0.26=pyha770c72_0
prompt_toolkit=3.0.26=hd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-opencv=4.5.5=py39hfa6204d_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.1=py39h5161555_0
pysocks=1.7.1=py39h2804cbe_4
python=3.9.10=hd16f9c5_2_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.2=cpu_py39hbfdb42d_0
pytorch-cpu=1.10.2=cpu_py39hd610c6a_0
pyzmq=22.3.0=py39h02c6a76_1
readline=8.1=hedafd6a_0
requests=2.27.1=pyhd8ed1ab_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=60.7.1=py39h2804cbe_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
sqlite=3.37.0=h72a2b83_0
stack_data=0.1.4=pyhd8ed1ab_0
terminado=0.13.1=py39h2804cbe_0
testpath=0.5.0=pyhd8ed1ab_0
tk=8.6.11=he1e0b03_1
tomli=2.0.0=pyhd8ed1ab_1
torchvision=0.10.0=py39h0a40b5a_0_cpu
tornado=6.1=py39h5161555_2
traitlets=5.1.1=pyhd8ed1ab_0
typed-ast=1.5.2=py39h5161555_0
typing_extensions=4.0.1=pyha770c72_0
tzdata=2021e=he74cb21_0
urllib3=1.26.8=pyhd8ed1ab_1
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.5.2=py39h2804cbe_1
x264=1!161.3030=h3422bc3_1
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
zeromq=4.3.4=hbdafb3b_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=hee7b306_1013
zstd=1.5.2=h861e0a7_0
